requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.2
exceptiongroup==1.1.3
h11==0.14.0
html5lib==1.1
idna==3.4
lxml==4.9.3
outcome==1.3.0.post0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.15.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.22.2
trio-websocket==0.11.1
urllib3==2.0.7
webencodings==0.5.1
wsproto==1.2.0
